requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
PyMySQL==1.1.0
cryptography>=46.0.0
python-barcode==0.15.1
reportlab==4.0.7
Pillow==10.1.0
python-decouple==3.8
